Philanthropic passing tests with postman image in README

+++ src/main/resources/db/create.sql
@@ -4,17 +4,17 @@ CREATE TABLE IF NOT EXISTS philanthropics (
  id int PRIMARY KEY auto_increment,
  name VARCHAR,
  type VARCHAR,
- latitude FLOAT,
- longitude FLOAT,
+ latitude DOUBLE,
+ longitude DOUBLE,
  charities VARCHAR,
- annualContribution FLOAT
+ annualContribution DOUBLE
 );
 
 CREATE TABLE IF NOT EXISTS socialenterprises (
  id int PRIMARY KEY auto_increment,
  name VARCHAR,
  type VARCHAR,
- latitude FLOAT,
- longitude FLOAT
+ latitude DOUBLE,
+ longitude DOUBLE
 );
